PN-4506: refactoring joining AR and RS tests and preparation for discoveredAddress check
